requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.30.1